Implement Simulation.save as a save_as method.
If Simulation.save receives an optional file_address parameter,
then the method works as a 'save_as' method now.
It changes the instances file_address property and then saves the instance.
This seems like a more natural and expected behaviour.
It enables the user to easily open an old simulation,
save it as another simulation,
change a few parameters and then run the calculation,
without the risk of overwriting the old simulation file.