Hotfix: Species.edit specimen
Species.edit_Specimen(item) did not work, if it recieved a string.
Due to the call:
	self.collision_rate_matrix[item, :]
Also if item is a string an the name is changed,
then the old item does not exist anymore, this leads to errors.
Now the index is generated in the beginning and the issue is fixed.