Completely rework the OutputFunction class
Move OutputFunction class into a function call.
Implement output_function(simulation, hdf5_group)
which returns the desired function.

Everything else needed to be changed in here.
A few noncritical bugs are found but only commented, not fixed.
This is necessary to keep the tests running but will be fixed soon.
The parameters of the subfunctions will be changed
when introducing the data class.

It is now possible to specifiy the place of the computation results
in the hdf5_file.
For this all the lower level functions receive a hdf5_group parameter.