Add out put "Complete_Distribution"
Complete Distribution writes the complete velocity distribution to the hdf5 file.
Currently, this is only for testing,
but it could be used later on in spatial homogeneous simulations.
So far no animation is allowed for this outputm as it is unclear how to do so.

Add test_case that checks this output.