Create Data class and rework computation.
A Data instance contains all parameters for the computation.
It will be the sole parameter for any of its sub functions:
	operator_splitting
	transport
	collision
	any output function
This is the first approach in pure python.
Soon, this will be moved into a Cython class.

Split up the Computation class into separate functions.
As stated above, the data instance shall be their only parameter.