Add Scheme class.
The Scheme class will handle the setup and storage
of any parameters related to computation schemes.
The available options can be configured by the constants:
	SUPP_SCHEME_VALUES
	REQ_SCHEME_PARAMETERS

Add load and save methods.
All Scheme parameters are saved as attributes of the "Computation" hdf5.Group.
Vice versa, every "Computation" attribute is also a Scheme parameter.

Add Scheme.__str__ method.
The Scheme class is essentially a dictionary.
Printing the Scheme prints all items in alphabetical order.

Add Scheme.__repr__ and Scheme.__eq__ methods.
The constants (see above) must follow a few rules.
Essentially Each key must be a valid variable name in python.
Thus:
	Keys contain no Spaces-> Use CamelCase instead.
	Use _ to separate parameter hierarchies.
This allows us to use eval(Scheme.__repr__)
to construct an identical copy of the instance from itself.
This is supposed to be a template for the assistants.
When every class, so even the Simulation class, has a valid __repr__,
we can easily edit the test cases, if it's  necessary.

Adjust computations switches to Scheme items.
	Remove coll_substeps attribute from Simulation.
	If necessary, we can implement this again, as a multistep ODE scheme.