Add script to test_data/generate_test_cases
The script constructs a dictionary of functions.
The key of any function is eqal to the file name of a hdf5 file.
This file can be created by calling the function.
The dictionary can be used for simple testing,
or, more importantly, editing the test cases.
This is a necessary function, as future commits might change attributes
or parameters. It is necessary to swiftly adjust the test cases to such changes.

By comparing their __str__ results the test cases can be checked for equality.

The current implementation is only a draft.
This should be implemented in a more pythonic way:
	The functions are redundant and ugly.
	Some redundant parts can probably be moved into a decorator.
	Reusing the existing functions, and ajusting after that,
	should help reduce further redundance.
	Putting each funtion into the dictionary, should be simpler.