Rework folder structure and parametrize tests.
All functions used in the computation process are moved
into a separate computation submodule. This includes the Data class.

All __init__.py files are emptied as much as possible.
Now, importing boltzpy gives direct access to all classes in the module.
This does NOT apply to the the Data class.
To access computation features the computation submodule must be imported.

Add test for collision_relations and use pytest.mark.parametrize.

Import abbreviations are changed:
	bp	for boltzpy
	bp_c	for boltzpy.constants
	cp	for boltzpy.computation
	bp_h	for boltzpy.helpers
	bp_ani	for animation