Add is_configured and is_set_up properties.
Add said properties to the classes:
	Simulation
	Grid
	SVGrid
	Species
	Scheme
This will be used later, to adjust a simulation to changing parameters.
eg. adding a specimen after setting up the velocity grid,
should either clear the grid or create it once again with the new specimen.

Add masses attribute to SVgrid.
This is an important parameter for the class.
Later the check_integrity methods will receive an optional parameter: simulation.
This will make comparisons with the other subclasses possible.