Bugfix: loading Scheme raises NotImplementedError
When loading a saved Scheme,
all values that were previously of type int
are returned as np.int64 by h5py.
When creating the repr, both
	isinstance(val, int)
	type(val) == int
returned False.
Thus, a NotImplementedError is raised.

Now, the Scheme.load() method converts np.int64's to int.

Change Scheme.keys() and Scheme.items() to return a dictionary view,
instead of a list. The second is ordered, but the first is not.
Since the list was created by the dict_view, this lead to
scheme.keys() == scheme.keys() being false,
since the lists are differently ordered.