Bugfix: Sphinx showed :obj:`...` at every Returns section.
Now all properties are documented in the following way:
	"""DATATYPE :
    	Description
    	"""
All methods use a Returns section.
Note that the return variable needs to have a name now
The Returns section looks like thos now:
   	"""DESCRIPTION OF THE METHOD

	Parameters
	----------
	param : DATATYPE

    	Returns
    	-------
    	VAR_NAME : DATATYE
    		IF NECESSARY A DESCRIPTION
    	"""

Sphinx also showed a warning due to the indentation in the class description of Collisions.
The line break/indent was in a math environment and is removed now.

Add intersphinx references for h5py
	h5py.Group -> h5py:Group
Those must have been missed previously.