Add TestCase class and add Geometry class.
TestCase class is a step to automatically update the test cases.
It is necessary to provide script-like test cases.
The class is an extended dictionary.
It is planned to create Simulations with Simulation(**testcase)
The class contains methods to save simulation data and update them.

Geometry class contains the initialization rules
and in the future calculation behaviour.
It is planned to merge it with the p_grid into
a class that contains all spatial data.