Change most of the new properties back into attributes, for performance.
Most are fine, however i_vels and vels are heavily used for computation.
This way is faster. However if every rule has a model attribute, ist consumes storage.
This can be circumvented, by creating a base class without the properties
and two child classes, one with properties and one with attributes.
But this is bad style and tedious to maintain.

As long as performance is not tight, this should be okay.